+++ pre_commit_bigfix/__init__.py
@@ -1,3 +1,3 @@
 """To support running this directly as a python module."""
 
-__version__ = "0.3.3"
+__version__ = "0.4.1"